Fixed Docker Compose Issue for Mobile
In my DockerFile I was using the RUN command
instead of the CMD command to start my application.
The slight distinction is the RUN command is used to
execute commands inside your image and are done at run time.
With every RUN command creating a new layer of your docker img.
The CMD command is used once the container starts.